myproject/backend/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
flask-cors==4.0.0
